remove Standalone platform
Standalone is not updated and onsolete. Strarting from now there will be just the App version. All the scripts will be tweaked to run as standalone scripts. This will increase the quality of the product and will allow a better use of energies by our team!
PS: This operation will not reduce the speed at all.