update: fix several issue and expand functionality
This is a major update: in these months I worked a lot to implement 2 possible models (Model A and Model B), i.e., 2 different types of neural networks. Now a lot of new tools are added and different model interpretation methods are implemented. I tried also to standardize the 2 approaches. Several compatibility issues are fixed and some modules were tested with linux. The update is still in progress, some less useful script may be broken.